ai-image-detection/requirements.txt
torch
transformers
flask
pillow
gunicorn
